# docs/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cybertech E-Commerce API
  description: |
    Backend REST API for the Cybertech e-commerce platform.

    Covers the full e-commerce lifecycle: product catalog, search,
    cart, wishlist, orders, payments (Stripe), reviews and moderation,
    bank cards (PCI-aware), discount campaigns, user management
    (Keycloak-backed) and behavioural event ingestion.

    Authentication is performed via Keycloak-issued JWT bearer tokens.
    Public endpoints (catalog browsing, registration, Stripe webhook)
    are explicitly whitelisted; every other endpoint requires a
    valid Bearer token and may further restrict to USER or ADMIN roles.
  contact:
    name: Cybertech Team
    email: dev@cybertech.local
  license:
    name: Proprietary
    url: https://cybertech.local/license
  version: "1.0"
servers:
- url: http://localhost:8081
  description: Local dev
- url: https://api.cybertech.local
  description: Production
security:
- keycloak: []
...
paths:
...
  /api/v1/webhooks/stripe:
    post:
      tags:
      - Stripe Webhook
      summary: Stripe Webhook endpoint
      description: Receives Stripe webhook events for payments and refunds. Returns
        200 on accepted/non-retriable; 400 on invalid signature or malformed payload.
      operationId: handleStripeEvent
      parameters:
      - name: Stripe-Signature
        in: header
        required: true
        schema:
          type: string
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: Webhook accepted (or non-retriable downstream failure ACKed
            to stop Stripe retries)
        "400":
          description: Invalid Stripe signature or malformed event payload
...
  /api/v1/services/user/register:
    post:
      tags:
      - User
      summary: Public signup
      description: |
        Anonymous endpoint (whitelisted in SecurityConfig#PUBLIC_URLS). Creates the Keycloak
        account and the local user in one saga; the optional nested bank card is stored
        encrypted. The response only carries id, email and username — the Keycloak id is
        deliberately not exposed to an anonymous caller.
      operationId: register
      parameters:
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        description: Signup payload
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequestDto"
        required: true
      responses:
        "201":
          description: "User created — body: {id, email, username}"
          content:
            application/json: {}
        "400":
          description: Invalid input data / Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "409":
          description: A user with this email / username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
...
  /api/v1/services/user/me:
    patch:
      tags:
      - User
      summary: Self-service profile update for the authenticated user
      description: |
        Frontend-gap #3 — patches the authenticated user's own profile (first/last name,
        phone number, address). The user UUID is resolved from the JWT subject so there
        is no IDOR risk: callers can only mutate their own row. Email/username are managed
        by Keycloak and intentionally NOT exposed here. Admin-only fields (status, role)
        are also off-limits to self-service.
      operationId: updateMe
      parameters:
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        description: Self-update payload — every field is optional; nulls leave the
          corresponding column untouched.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSelfUpdateRequestDto"
        required: true
      responses:
        "200":
          description: User profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "400":
          description: Invalid input data / Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: User not found (JWT subject points to a stale Keycloak entry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
...
  /api/v1/services/user/ok:
    get:
      tags:
      - User
      summary: Health check endpoint
      description: A simple endpoint to check if the UserController is responsive.
      operationId: healthCheck
      parameters:
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      responses:
        "200":
          description: Service is up and running
          content:
            application/json:
              schema:
                type: string
                example: Hello Guys !!!
  /api/v1/services/user/get/{userUuid}:
    get:
      tags:
      - User
      summary: Request a User by UUID
      description: Fetches a user's details based on their unique UUID. Non-admin
        callers may only read their own profile.
      operationId: getUserByUuid
      parameters:
      - name: userUuid
        in: path
        description: UUID for searching a user
        required: true
        schema:
          type: string
          format: uuid
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
          - 1.0.0
      responses:
        "200":
          description: User found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "400":
          description: "Bad request (e.g., invalid UUID format)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Operation forbidden - non-admin caller fetching another user's
            profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
...
components:
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        message:
          type: string
        httpStatusCode:
          type: integer
          format: int32
        errorCodeType:
          type: string
          enum:
          - TECHNICAL
          - FUNCTIONAL
...
    BankCardCreationRequestDto:
      type: object
      properties:
        cardHolderName:
          type: string
          minLength: 1
        cardNumber:
          type: string
          maxLength: 19
          minLength: 13
          pattern: "\\d{13,19}"
        expiryDate:
          type: string
          minLength: 1
          pattern: "(0[1-9]|1[0-2])/[0-9]{4}"
        cardType:
          type: string
          enum:
          - VISA
          - MASTERCARD
          - AMERICAN_EXPRESS
          - DISCOVER
          - OTHER
      required:
      - cardHolderName
      - cardNumber
      - cardType
      - expiryDate
    UserCreateRequestDto:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        firstName:
          type: string
          maxLength: 50
          minLength: 2
        lastName:
          type: string
          maxLength: 50
          minLength: 2
        sex:
          type: string
          enum:
          - M
          - F
        phoneNumber:
          type: string
          minLength: 1
        favoriteCommunicationChanel:
          type: string
          enum:
          - EMAIL
          - SMS
          - PUSH_NOTIFICATION
        street:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        zipCode:
          type: string
          minLength: 1
        country:
          type: string
          minLength: 1
        birthDate:
          type: string
          format: date-time
        password:
          type: string
          maxLength: 2147483647
          minLength: 8
        bankCardCreationRequestDto:
          $ref: "#/components/schemas/BankCardCreationRequestDto"
      required:
      - city
      - country
      - email
      - favoriteCommunicationChanel
      - firstName
      - lastName
      - password
      - phoneNumber
      - sex
      - street
      - zipCode
...
    UserResponseDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        sex:
          type: string
          enum:
          - M
          - F
        address:
          type: string
        birthDate:
          type: string
          format: date-time
        role:
          type: string
          enum:
          - USER
          - ADMIN
...
    UserSelfUpdateRequestDto:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 50
          minLength: 2
        lastName:
          type: string
          maxLength: 50
          minLength: 2
        phoneNumber:
          type: string
        street:
          type: string
        city:
          type: string
        zipCode:
          type: string
        country:
          type: string
...
  securitySchemes:
    keycloak:
      type: http
      description: "Keycloak-issued JWT access token. Obtain one from the realm token\
        \ endpoint and pass it as the Authorization header: 'Bearer <token>'."
      scheme: bearer
      bearerFormat: JWT